Union selector constant into tables
This will union selector constant into tables.
