+++ scripts/boj/constants/index.js
@@ -1,7 +1,6 @@
 import { LANGUAGES } from "./languages.js";
 import { PLATFORMS } from "./platforms.js";
-import { SELECTORS } from "./selectors.js";
-import { STATUS_TABLE_INDICES } from "./tables.js";
+import { SELECTORS, STATUS_TABLE_INDICES } from "./tables.js";
 import { DARK_THEMES } from "./themes.js";
 
 export { LANGUAGES, PLATFORMS, SELECTORS, STATUS_TABLE_INDICES, DARK_THEMES };
