--- scripts/boj/constants/selectors.js
@@ -1,8 +1,0 @@
-export const SELECTORS = {
-  LOGIN_ID: '.username',
-  SOURCE_TEXTAREA: "textarea[name='source']",
-  STATUS_TABLE: '#status-table',
-  CODE_MIRROR: '.CodeMirror',
-  PROBLEM_LINK: "table a[href^='/problem/']",
-  RESULT_TABLE: 'table tbody tr td',
-};
