+++ scripts/boj/constants/tables.js
@@ -5,3 +5,12 @@ export const STATUS_TABLE_INDICES = {
   RESULT: 3,
   LANGUAGE: 6,
 };
+
+export const SELECTORS = {
+  LOGIN_ID: '.username',
+  SOURCE_TEXTAREA: "textarea[name='source']",
+  STATUS_TABLE: '#status-table',
+  CODE_MIRROR: '.CodeMirror',
+  PROBLEM_LINK: "table a[href^='/problem/']",
+  RESULT_TABLE: 'table tbody tr td',
+};
